change theme functionality completed
