+++ src/game.py
@@ -2,7 +2,7 @@ import pygame
 from const import *
 from board import Board
 from dragger import Dragger
-
+from config import Config
 """
 Game Class: shows board and pieces on screen, initializes board and mouse dragger objects
 """
@@ -13,19 +13,20 @@ class Game:
         self.hovered_sqr = None
         self.board = Board()
         self.dragger = Dragger()
+        self.config = Config()
 
     def show_bg(self,surface): #surface will be self.screen from main
+        theme = self.config.theme
+
+
         #draw board
         for row in range(ROWS):
             for col in range(COLS):
-                if(row+col) % 2 == 0:
-                    color = (234,235,200) #beige
-
-                else:
-                    color = (119,154,88) #dark green
-
+                #color
+                color = theme.bg.light if (row+col)%2==0 else theme.bg.dark
+                #rect
                 rect = (col * SQSIZE, row * SQSIZE, SQSIZE, SQSIZE)
-
+                #blit
                 pygame.draw.rect(surface, color, rect)
 
 
@@ -49,13 +50,15 @@ class Game:
 
     #show moves of the piece that the user clicks on
     def show_moves(self, surface):
+        theme = self.config.theme
+
         if self.dragger.dragging:
             piece = self.dragger.piece
 
             #loop through all valid moves and blit them 
             for move in piece.moves:
                 # color
-                color = '#C86464' if(move.final.row + move.final.col) % 2 == 0 else '#C84646'
+                color = theme.moves.light if(move.final.row + move.final.col) % 2 == 0 else theme.moves.dark
                 #rect
                 rect = (move.final.col * SQSIZE, move.final.row * SQSIZE, SQSIZE, SQSIZE)
                 
@@ -63,12 +66,14 @@ class Game:
                 pygame.draw.rect(surface, color, rect)
 
     def show_last_move(self,surface):
+        theme = self.config.theme
+
         if self.board.last_move:
             initial = self.board.last_move.initial
             final = self.board.last_move.final
 
             for pos in [initial,final]:
-                color = (244, 247, 116) if (pos.row + pos.col) % 2 == 0 else (172, 195, 51)
+                color = theme.trace.light if (pos.row + pos.col) % 2 == 0 else theme.trace.dark
 
                 rect = (pos.col* SQSIZE, pos.row * SQSIZE , SQSIZE, SQSIZE)
 
@@ -88,3 +93,8 @@ class Game:
                 
     def next_turn(self):
         self.next_player = 'white' if self.next_player == 'black' else 'black'
+
+
+    def change_theme(self):
+        self.config.change_theme()
+
